+++ dialector/migrator.go
@@ -7,7 +7,7 @@ import (
 	"reflect"
 	"strings"
 
-	"github.com/yourname/gorm-iotdb/internal/model"
+	"github.com/wkk778/gorm-iotdb/internal/model"
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
 	"gorm.io/gorm/schema"