+++ tests/iotdb_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 	"time"
 
-	gormiotdb "github.com/yourname/gorm-iotdb"
+	gormiotdb "github.com/wkk778/gorm-iotdb"
 	"gorm.io/gorm"
 )